:beers: Resolve class conflict
Issue: #27

Am redenumit clasa CrawTask in CrawlTask si am eliminat dublura
